Added index.html and manchurian.html
Created the index.html file which will be the homepage of our recipes website.
Created a folder inside the parent folder called recipes.
Created a manchurian.html file in the recipes folder.
Added basic HTML boilerplate in all the html files.
Added headings and titles to both the html files.
Linked the manchurian.html file to the index.html file.